Use two large `do` forms for the whole file
Lua has a limit of 200 locals per function (including the top-level
'main' function), and this file exceeds that. The previous solution was
to wrap small sections in `do` to exclude those sections from the
top-level locals count. This worked fine, but since I was doing only a
few functions at a time, it didn't lower the locals count very much.
This solution makes it so that the file is basically split in half, so
eliminates many more locals in one fell swoop.

Additionally, I was using `(lua :do) ... (lua :end)` to eliminate issues
with indentation, but the fennel compiler doesn't know about that, and
wasn't able to warn about unknown locals at compile time. This commit
falls back to the usual `(do ...)` form.