Abandon ns macro, move declare and defn
General consensus in #fennel is that trying to abstract around the lua
module system is a bad move. For a module this big I still find it handy
to have a macro that defines a local and sticks it in the module table.
Plus I need some forward declarations, and I think `(declare x)`
indicates intent better than `(var x nil)`.

The encouraged fennel style is to return a map of functions at the end
of the module, rather than adding them to a module table along the way,
but that won't work w/ the two big do forms anyways.