Remove key-opying from deep-copy; add deep-copy-with-keys
Copying keys is tricky, since keys are generally used for hash table
lookup and nothing else. When tables are used _as_ keys, the address is
used for hashing, so two keys that are equal in value will never hash to
the same key. So in general, there's rarely a reason to make a copy of
the key, even if you want a deep copy of the values.

This change lets you opt in to copying keys in the rare case where you
might want to.